actually got rid of dump_arm_movement_time, and made current_action_state uint32_t to not cause issues
ALSO REMOVED cv_bridge BECAUSE IT HAS BEEN PISSING ME OFF

+++ src/network_communication/src/net_node.cpp
@@ -3,7 +3,6 @@
 #include <memory>
 #include <std_msgs/msg/detail/u_int8__struct.hpp>
 #include <string>
-#include <cv_bridge/cv_bridge.h>
 #include <sensor_msgs/msg/image.hpp>
 #include "std_msgs/msg/string.hpp"
 #include "std_msgs/msg/u_int8.hpp"
@@ -73,7 +72,7 @@ std::vector<std::pair<std::string, StickFieldPtr>> stickFields = {
 //const char* CONTROL_STATION_IP_CLIENT = std::getenv("CONTROL_STATION_IP");
 ThreadInfo info;
 int counter = 0;
-uint8_t current_action_state = 0;
+uint32_t current_action_state = 0;
 
 const int MOTOR_CURRENT_BYTES = 4;
 int Socket(ThreadInfo *info)
